Fix alignment and offset of bit-fields
There are two use cases for getting alignment of a bit-field.
1. Computing alignment of containing struct;
2. Computing offsets and size of containing struct.

In first case, alignment attribute is applied as for a usual
field (can only increase alignment unless in packed struct).
Zero-length bit-filds are ignored while computing struct alignment.

In second case, alignment attribute is effective only if it
overaligns the type of the bit-field, or if the bit-field is not
zero-length and fits into alignment unit, or if the bit-field is
not zero-length and struct is packed.