Create new branch for CFA mutator
The goal is to mutate input CFA in such a way, that the reuslt of
analysis -- specifically, an exception thrown during analysis --
remains the same. This way a big input program which crashes
CPAchecker can be minimized automatically.

The work is based on Delta Debugging algorithm. This is refactoring
of https://github.com/mutilin/cpachecker/tree/cfamutation.