Change option for checks after rollback
If there is a problem with mutations, such that rollback can not
restore dought-for bug occurance, every following round will be
rollbacked too. So instead of checking every rollback, only
suspiciously many rollbacks in a row must be checked. On the other
hand, we want to find it as soon as possible.

So make option for how many rollbacks in a row are suspicious.