Remove nodes that are reachable via summary edges too
If a node is after summary edge of a non-returning function, it is
reachable only via that summary edge (there is no return edge). In
this case every node that is reachachable from this node only is
removed, but the node that is reachable via possible leaving summary
edge was not considered.