Fix bug for ONLY_COMPLEMENTS part to remove
In often case of two deltas, algorithm tries to remove the second
one as a complement. If test passes, the first delta is marked as
safe, and the second delta is restored. The bug is that the algorithm
tries to remove 'the first delta' while it is already marked as safe,
i.e. removes 0 elements.

For this case split deltas if iterator has no next delta or there
is only one delta remaining.