Add CHECK_WHOLE stage so DD*Max checks whole before repeating
If the first found cause is the only cause, dd*max will remove it
in the first pass, return removed, and check all but the cause.
As test passes, dd*max found maximum.

If dd*max can not check whole, it assumes there are other causes,
so it will 'find' them and remove some elements that can safely
remain in maximum.